requirements.txt
numpy
scipy
sounddevice
matplotlib
pyrtlsdr
pyModeS
